flame: overlay: Copy some refresh rate config overlays from Pixel 6 (oriole)
Pixel 4/XL has this very annoying flicker when ambient is near
to close to this threshold which forces the refresh rate to go
from peak (90Hz) to lowest possible (60Hz). Pixel 6, fortunately,
does not exhibit this behavior. Since it's also 90Hz, we can safely
copy these configs and get a much better UX.